+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+## [0.1.7] 
+- New dashboard
+- Export table to CSV
+- Fix storage query error when table missed `price_per_month` column
+
+## [0.1.6] 
 - Notify when new Finala version release
 
 ## [0.1.5]